chore(bupkis): add property test fuzzing infrastructure
Adds a fuzzer that runs property tests with 1 million iterations each,
distributing work across CPU cores via a child process pool.

Key changes:
- Add fuzz.ts orchestrator and fuzz-worker.ts for parallel execution
- Extract testConfigs from test files into configs/ directory to avoid
  importing node:test when loading fuzz targets
- Add fuzzing-registry.ts to consolidate all fuzz targets
- Add test:fuzz script (not run in CI)
