+++ package.json
@@ -31,6 +31,7 @@
     "test": "npm run test --workspaces",
     "test:coverage": "npm run test:coverage --workspaces --if-present",
     "test:dev": "npm run test:dev --workspaces --if-present",
+    "test:fuzz": "npm run test:fuzz --workspace=bupkis",
     "test:node20": "npm run test:node20 --workspaces",
     "test:property": "npm run test:property --workspaces --if-present",
     "test:snapshots": "npm run test:snapshots --workspaces --if-present",
